Check for existing DOIs
This extension will complain if it finds objects with the DOI given when
* ingesting an object via DOI-import
* creating a new object and manually insert an existing DOI
* editing an object with a DOI that exists elsewhere

It should find all objects in the same namespace as the object under validation, or in the same namespace as the collection the object under validation is in. It should not complain on object with unrelated namespace (both of the object and the collections it is in).

Finally, this extension provides a simplistic form in which one can paste several DOIs and will give links to existing objects (useful for batch-ingests).